refactor(parsing): pattern matcher 클래스 추상화 및 변수이름 축약
CustomDelimiterParser와 RegexPatternMatcher와의 결합도를 낮췄습니다. 그리고 일부 길이기 긴 변수명을 축약형으로 사용하였습니다.
