+++ src/main/java/calculator/pattern/PatternMatcher.java
@@ -1,0 +1,7 @@
+package calculator.pattern;
+
+public interface PatternMatcher {
+    int countMatches(String input);
+
+    String getMatch(String input);
+}
