+++ src/main/java/calculator/pattern/RegexPatternMatcher.java
@@ -1,14 +1,24 @@
-package calculator.regex;
+package calculator.pattern;
 
 import java.util.regex.Pattern;
 
-public class RegexPatternMatcher {
+public class RegexPatternMatcher implements PatternMatcher {
     private final Pattern pattern;
 
     public RegexPatternMatcher(String regex) {
         this.pattern = Pattern.compile(regex);
     }
 
+    @Override
+    public String getMatch(String input) {
+        var matcher = pattern.matcher(input);
+        if (matcher.find()) {
+            return matcher.group(1);
+        }
+        return null;
+    }
+
+    @Override
     public int countMatches(String input) {
         var matcher = pattern.matcher(input);
 
@@ -17,13 +27,5 @@ public class RegexPatternMatcher {
             count++;
         }
         return count;
-    }
-
-    public String getMatch(String input) {
-        var matcher = pattern.matcher(input);
-        if (matcher.find()) {
-            return matcher.group(1);
-        }
-        return null;
     }
 }
